backend/requirements.txt
# Enterprise Insights Copilot - Backend Requirements

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
openpyxl==3.1.2
pyarrow==14.0.1

# AI/ML Framework
langchain>=0.1.0
langgraph>=0.0.62
langsmith>=0.0.70
openai>=1.3.7
tiktoken>=0.5.2

# Vector Database
pinecone-client==2.2.4
sentence-transformers==2.2.2

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7
rich==13.7.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
